Colocando a criptografia no utils
E cleanup
